Add initial helper scripts
This adds two new helper scripts that are useful when working with large
comparisons.
scripts/gen_constants.py creates a constants file that can help reduce
the number of constants generated in a comparison.
scripts/clean_csv.py performs a number of simplifications of a csv file
and creates a new output that is much easier for a human to review. One
important note here is that the resulting report is not a valid csv
file.